add modules to command constant file
