+++ src/extension/src/constants/commands.ts
@@ -30,3 +30,15 @@ export enum ExtensionCommand {
   GET_TEMPLATE_INFO = "get-template-info",
   GET_DEPENDENCY_INFO = "check-dependency"
 }
+
+export enum ExtensionModule {
+  Azure = "Azure",
+  Generate = "GenerateExperience",
+  Telemetry = "Telemetry",
+  Validator = "Validator",
+  VSCodeUI = "VSCodeUI",
+  Logger = "Logger",
+  DependencyChecker = "DependencyChecker",
+  CoreTSModule = "CoreTSModule",
+  Defaults = "Defaults"
+}
