--- src/extension/src/constants/modules.ts
@@ -1,11 +1,0 @@
-export enum ExtensionModule {
-  Azure = "Azure",
-  Generate = "GenerateExperience",
-  Telemetry = "Telemetry",
-  Validator = "Validator",
-  VSCodeUI = "VSCodeUI",
-  Logger = "Logger",
-  DependencyChecker = "DependencyChecker",
-  CoreTSModule = "CoreTSModule",
-  Defaults = "Defaults"
-}
